+++ program/Hardware/spwm.h
@@ -13,6 +13,6 @@ extern uint16_t spwm_list_size;
 
 void get_spwm_list(uint32_t pwm_freq, uint16_t pwm_duty, float spwm_freq);
 void tim1_reset(uint32_t prescaler, uint32_t period);
-uint8_t spwm_set(float spwm_freq);
+uint8_t spwm_set(uint16_t spwm_freq);
 
 #endif //PROGRAM_SPWM_H
